initial commit of ttm18 shiny test
This is an attempt to provide an interactive version of the travelTimeComparison() function I have implemented in py. So far each new map calculation takes about five minutes, so not that feasible. To be fair, this can't be deployed to shinyapps.io because it needs the full TTM18 to work (13,8 GB of data)