messy version: fst recalculation of traveltime df
save this test version of the script for posterity. Extensive testing was made with data.table::fread, vroom, parquet and fst to make recalculation of dataframe faster. I will be moving forward with fst, which unfortunately needs a conversion of files from txt-->fst, but that seems to drop load times from 2-3 minutes to mere seconds.